Chunking Visualization
Added a library to help visualise chunks after indexing. A html file will be created when index is run with the flag visualise that can be used to see the chunks

A script to evaluate the retrieval for 5 questions based on gold text for different chunking strategies is included. The result is stored in an html/json.